refactor: Member 패키지 하위 파일들을 Auth 패키지로 이동
