+++ src/main/java/com/github/bbooong/bangumall/auth/dto/MemberCreateRequest.java
@@ -1,3 +1,3 @@
-package com.github.bbooong.bangumall.member.dto;
+package com.github.bbooong.bangumall.auth.dto;
 
 public record MemberCreateRequest(String email, String password) {}
